refactor: Move cleanLogContent to model/backend/gitlab and add tests
